[Refactor] refactoring codes
